Abstract code for running test in parallel
